started (finely) the final attitude towards batch running
The basic revalation after a week with rosen@technion.ac.il is that parralelizing has its limits. Therefor it will always make sense to submit different cases in parallel. that is - not one 150 core job, but 10 15 core jobs. For me - running several directions of wind at a given site - this makes perfect sense. This was done on the hanan220 machine instead on hanan@rosen because the paraview was down and i wanted to also make automatic png files of surface cuts to see solutions for my current testZ0Influence runs on 2d hills which look like shit as aviv commented.
So - the last merges where problematic to say the list.... now - on to work on rosen and combine the _noloop version of the 2D hill run, in the new prepareCase_2dHill.py and processCase_2dHill.py, under the (original) testZ0Influence.py.